+++ CHANGELOG.rst
@@ -40,6 +40,9 @@ Fixed
 
 - Don't make yaml formatting ugly during config upload.
 
+- "No value provided for non-null ReplicaStatus" GraphQL error
+  after removing a replica from the `box.space._cluster`.
+
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Enhanced is WebUI
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~